third_party/blink/web_tests/flag-specific/highdpi/external/wpt/css/css-align/baseline-rules/synthesized-baseline-grid-001-expected.txt

This is a testharness.js-based test.
PASS .wrapper 1
PASS .wrapper 2
FAIL .wrapper 3 assert_equals: 
<div class="wrapper">
  <canvas></canvas>
  <div style="display: inline-block;" class="magenta-dotted-border" data-offset-y="0">
    <div style="display: grid;" class="border-padding-margin" data-offset-y="25"></div>
  </div>
</div>
offsetTop expected 0 but got 1
PASS .wrapper 4
PASS .wrapper 5
PASS .wrapper 6
PASS .wrapper 7
Harness: the test ran to completion.
